docker/pydata/pythonProject/requirements.txt
cryptography==3.4.8
kafka-python==2.0.2
